big refactor
- Separated accounting from daytrade calculations
- Made the propagation of changes from adding a transaction be with a "sudo observer"
- Big fat refactor (probably not done still)
