+++ gerenciador-investimentos/src/main/kotlin/com/vitorbasso/gerenciadorinvestimentos/service/IAssetService.kt
@@ -3,17 +3,10 @@ package com.vitorbasso.gerenciadorinvestimentos.service
 import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Asset
 import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Stock
 import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Wallet
-import com.vitorbasso.gerenciadorinvestimentos.enum.TransactionType
-import java.math.BigDecimal
 
 interface IAssetService {
-    fun deleteAsset(walletId: Long, ticker: String)
 
-    fun addTransactionToAsset(
-        wallet: Wallet,
-        stock: Stock,
-        amount: Int,
-        cost: BigDecimal,
-        type: TransactionType
-    ) : Asset
+    fun getAsset(wallet: Wallet, stock: Stock): Asset
+
+    fun deleteAsset(walletId: Long, ticker: String)
 }
