+++ gerenciador-investimentos/src/main/kotlin/com/vitorbasso/gerenciadorinvestimentos/service/facade/TransactionServiceFacadeImpl.kt
@@ -1,14 +1,17 @@
 package com.vitorbasso.gerenciadorinvestimentos.service.facade
 
-import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Asset
 import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Transaction
+import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Wallet
 import com.vitorbasso.gerenciadorinvestimentos.dto.request.TransactionRequest
 import com.vitorbasso.gerenciadorinvestimentos.exception.CustomWrongDateException
+import com.vitorbasso.gerenciadorinvestimentos.service.IAssetService
+import com.vitorbasso.gerenciadorinvestimentos.service.IStockService
 import com.vitorbasso.gerenciadorinvestimentos.service.ITransactionService
-import com.vitorbasso.gerenciadorinvestimentos.service.concrete.StockService
+import com.vitorbasso.gerenciadorinvestimentos.service.IWalletService
+import com.vitorbasso.gerenciadorinvestimentos.service.concrete.AccountingService
 import com.vitorbasso.gerenciadorinvestimentos.service.concrete.TransactionService
-import com.vitorbasso.gerenciadorinvestimentos.util.AccountantUtil
 import com.vitorbasso.gerenciadorinvestimentos.util.SecurityContextUtil
+import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import java.time.DayOfWeek
@@ -18,24 +21,16 @@ import java.time.LocalDateTime
 @Service
 internal class TransactionServiceFacadeImpl(
     private val transactionService: TransactionService,
-    private val stockService: StockService,
-    private val assetService: AssetServiceFacadeImpl,
-    private val walletService: WalletServiceFacadeImpl,
-    private val monthlyWalletService: MonthlyWalletServiceFacadeImpl
+    private val stockService: IStockService,
+    private val assetService: IAssetService,
+    @Qualifier("walletServiceFacadeImpl")
+    private val walletService: IWalletService,
+    private val accountingService: AccountingService
 ) : ITransactionService {
 
     @Transactional
-    override fun performTransaction(transactionRequest: TransactionRequest) = this.assetService.addTransactionToAsset(
-        wallet = this.walletService.getWallet(transactionRequest.walletId),
-        stock = this.stockService.getStock(transactionRequest.ticker),
-        amount = transactionRequest.quantity,
-        cost = transactionRequest.value,
-        type = transactionRequest.type
-    ).let {
-        transactionRequest.getTransaction(it)
-    }.let {
-        processTransaction(it)
-    }
+    override fun performTransaction(transactionRequest: TransactionRequest) =
+        processTransaction(transactionRequest.getTransaction())
 
     @Transactional
     override fun deleteTransaction(transactionId: Long) {
@@ -45,27 +40,17 @@ internal class TransactionServiceFacadeImpl(
         )
         val transactions = this.transactionService.findTransactionsOnSameDate(transactionToDelete).toMutableList()
         transactions.remove(transactionToDelete)
-        this.transactionService.saveAll(AccountantUtil.processDaytrade(transactions))
+//        this.transactionService.saveAll(AccountantService.processDaytrade(transactions))
         this.transactionService.deleteTransaction(transactionToDelete)
     }
 
     private fun processTransaction(transaction: Transaction): Transaction {
-        val accountantReport = AccountantUtil.accountForNewTransaction(
+        val accountantReport = accountingService.account(
             transaction,
             this.transactionService.findFromOneBeforeTransactionDate(transaction)
         )
+
         this.transactionService.saveAll(accountantReport.transactionsReport)
-
-        this.walletService.processAccountantReport(
-            transaction.asset.wallet,
-            accountantReport,
-            this.monthlyWalletService
-        )
-
-        this.assetService.processAccountantReport(
-            transaction.asset,
-            accountantReport
-        )
 
         return accountantReport.transactionsReport.findLast { it.transactionDate.isEqual(transaction.transactionDate) }
             ?: transaction
@@ -76,11 +61,14 @@ internal class TransactionServiceFacadeImpl(
             (it.dayOfWeek != DayOfWeek.SATURDAY && it.dayOfWeek != DayOfWeek.SUNDAY)
     } ?: throw CustomWrongDateException()
 
-    private fun TransactionRequest.getTransaction(asset: Asset) = Transaction(
+    private fun TransactionRequest.getTransaction() = Transaction(
         type = this.type,
         quantity = this.quantity,
         value = this.value,
-        asset = asset,
+        asset = assetService.getAsset(
+            wallet = walletService.getWallet(this.walletId) as Wallet,
+            stock = stockService.getStock(this.ticker)
+        ),
         transactionDate = checkDate(this.date)
     )
