+++ gerenciador-investimentos/src/main/kotlin/com/vitorbasso/gerenciadorinvestimentos/service/facade/WalletServiceFacadeImpl.kt
@@ -1,14 +1,16 @@
 package com.vitorbasso.gerenciadorinvestimentos.service.facade
 
 import com.vitorbasso.gerenciadorinvestimentos.domain.IWallet
+import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Transaction
 import com.vitorbasso.gerenciadorinvestimentos.domain.concrete.Wallet
 import com.vitorbasso.gerenciadorinvestimentos.dto.request.WalletUpdateRequest
 import com.vitorbasso.gerenciadorinvestimentos.enum.ManagerErrorCode
 import com.vitorbasso.gerenciadorinvestimentos.exception.CustomBadRequestException
+import com.vitorbasso.gerenciadorinvestimentos.service.IAccountingServiceSubscriber
 import com.vitorbasso.gerenciadorinvestimentos.service.IWalletService
+import com.vitorbasso.gerenciadorinvestimentos.service.concrete.AccountingService
 import com.vitorbasso.gerenciadorinvestimentos.service.concrete.ClientService
 import com.vitorbasso.gerenciadorinvestimentos.service.concrete.WalletService
-import com.vitorbasso.gerenciadorinvestimentos.util.AccountantUtil
 import com.vitorbasso.gerenciadorinvestimentos.util.SecurityContextUtil
 import org.springframework.stereotype.Service
 import java.time.LocalDate
@@ -16,29 +18,29 @@ import java.time.LocalDate
 @Service
 internal class WalletServiceFacadeImpl(
     private val walletService: WalletService,
+    private val monthlyWalletService: MonthlyWalletServiceFacadeImpl,
     private val clientService: ClientService
-) : IWalletService {
+) : IWalletService, IAccountingServiceSubscriber {
 
-    override fun getWalletCollection()
-    = this.clientService.getClient(SecurityContextUtil.getClientDetails().id).wallet
+    override fun getWalletCollection() = this.clientService.getClient(SecurityContextUtil.getClientDetails().id).wallet
 
-    override fun getWallet(walletId: Long)
-    = this.walletService.getWallet(walletId, SecurityContextUtil.getClientDetails().id).validate()
+    override fun getWallet(walletId: Long) =
+        this.walletService.getWallet(walletId, SecurityContextUtil.getClientDetails().id).validate()
 
     override fun saveWallet(walletToSave: IWallet) = this.walletService.saveWallet(
         SecurityContextUtil.getClientDetails(),
         walletToSave as Wallet
     )
 
-    override fun updateWallet(walletId: Long, walletUpdateRequest: WalletUpdateRequest)
-    = this.walletService.updateWallet(
-        this.walletService.getWallet(
-            walletId = walletId,
-            clientId = SecurityContextUtil.getClientDetails().id,
-            exception = CustomBadRequestException(ManagerErrorCode.MANAGER_06)
-        ).validate(),
-        walletUpdateRequest
-    )
+    override fun updateWallet(walletId: Long, walletUpdateRequest: WalletUpdateRequest) =
+        this.walletService.updateWallet(
+            this.walletService.getWallet(
+                walletId = walletId,
+                clientId = SecurityContextUtil.getClientDetails().id,
+                exception = CustomBadRequestException(ManagerErrorCode.MANAGER_06)
+            ).validate(),
+            walletUpdateRequest
+        )
 
     override fun deleteWallet(walletId: Long) {
         this.walletService.deleteWallet(
@@ -50,23 +52,24 @@ internal class WalletServiceFacadeImpl(
         )
     }
 
-    fun processAccountantReport(
-        wallet: Wallet,
-        accountantReport: AccountantUtil.AccountantReport,
-        monthlyWalletService: MonthlyWalletServiceFacadeImpl
-    ) = wallet.validate().let { walletValidated ->
-        accountantReport.walletsReport.forEach {
-            this.walletService.processWalletReport(
-                walletValidated,
-                it.value,
-                it.key,
-                monthlyWalletService
-            )
+    override fun processAccountantReport(
+        transaction: Transaction,
+        accountantReport: AccountingService.AccountantReport
+    ): AccountingService.AccountantReport {
+        transaction.asset.wallet.validate().let { walletValidated ->
+            accountantReport.walletsReport.forEach {
+                this.walletService.processWalletReport(
+                    wallet = walletValidated,
+                    walletReport = it.value,
+                    walletMonth = it.key,
+                    monthlyWalletService = monthlyWalletService
+                )
+            }
         }
+        return accountantReport
     }
 
-    private fun isValid(wallet: Wallet)
-    = wallet.walletMonth.withDayOfMonth(1) == LocalDate.now().withDayOfMonth(1)
+    private fun isValid(wallet: Wallet) = wallet.walletMonth.withDayOfMonth(1) == LocalDate.now().withDayOfMonth(1)
 
     private fun Wallet.validate() = this.takeIf { isValid(it) } ?: walletService.enforceWalletMonth(this)
